distributions: add basic `DistributionManager::compute_issuance`
